Move logic in App component to 'GameSession' to get ready for Wlecome Screen (Issue #4)
